Intelligent frontend activation & better VS integration
- instead of blindly create frontend by a fixed class id, collector now attempts to bring up MSVS-integrated frontend, then, if failed - global registered frontend, if even that one fails, it falls back to 'null';
- initializer / library are now added relatively to MICROPROFILERDIR, which helps when upgrading profiler extension from VisualStudio (project contains no references to profiler installation directory - MICROPROFILERDIR does);
- installer changed to add installation directory to the Path via MICROPROFILERDIR;
- codebase refactored to better conform to C++ standard by globally enabling "Disable Language Extensions".

+++ micro-profiler/frontend/controls.cpp
@@ -65,7 +65,7 @@ namespace micro_profiler
 
 			virtual void resize(unsigned cx, unsigned cy, positioned_native_views &nviews)
 			{
-				view_location l = { 0, 0, cx, cy };
+				view_location l = { 0, 0, static_cast<int>(cx), static_cast<int>(cy) };
 				nviews.push_back(positioned_native_view(*this, l));
 			}
